Revamp how system is loaded into Emacs.
ede/ede-patches.el:
(ede-patches-type-of-slot): New fcn that queries or
sets  the :TYPE of a slot in a laoded object.
Fix the :type of the :rules slot on ede-proj-makefile
so that we can load Project files in Emacs 26.
(other files):
Adjust how things load so that extensions can
be loaded first thing in your .emacs file.